trying a new way of handling death

+++ src/main/java/me/realized/duels/data/MatchData.java
@@ -12,11 +12,11 @@ public class MatchData implements MatchInfo {
     @Getter
     private final long time;
     @Getter
-    private final int duration;
+    private final long duration;
     @Getter
     private final double health;
 
-    public MatchData(final String winner, final String loser, final long time, final int duration, final double health) {
+    public MatchData(final String winner, final String loser, final long time, final long duration, final double health) {
         this.winner = winner;
         this.loser = loser;
         this.time = time;
